[FEAT] : CommunityCategorySection 컴포넌트 분리 및 추상화
